chore: Move function to load model from MLflow to utils file
